refactor; cleanup double use
