+++ tests/test_unit/test_cdf_tk/test_commands/test_build.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-from collections.abc import Iterable
 from pathlib import Path
 
 import pytest
 from _pytest.monkeypatch import MonkeyPatch
-from cognite.client.data_classes.data_modeling import DataModelId
 
 from cognite_toolkit._cdf_tk.commands.build import BuildCommand
 from cognite_toolkit._cdf_tk.data_classes import Environment
@@ -14,7 +12,7 @@ from cognite_toolkit._cdf_tk.exceptions import (
     ToolkitMissingModuleError,
 )
 from cognite_toolkit._cdf_tk.hints import ModuleDefinition
-from cognite_toolkit._cdf_tk.loaders import DataModelLoader, TransformationLoader
+from cognite_toolkit._cdf_tk.loaders import TransformationLoader
 from cognite_toolkit._cdf_tk.tk_warnings import LowSeverityWarning
 from tests import data
 
@@ -91,78 +89,3 @@ class TestBuildCommand:
         assert len(transformation_files) == 2
         sql_files = [f for f in (tmp_path / "transformations").iterdir() if f.is_file() and f.suffix == ".sql"]
         assert len(sql_files) == 2
-
-
-def valid_yaml_semantics_test_cases() -> Iterable[pytest.ParameterSet]:
-    yield pytest.param(
-        """
-- dbName: src:005:test:rawdb:state
-- dbName: src:002:weather:rawdb:state
-- dbName: uc:001:demand:rawdb:state
-- dbName: in:all:rawdb:state
-- dbName: src:001:sap:rawdb
-""",
-        Path("build/raw/raw.yaml"),
-        id="Multiple Raw Databases",
-    )
-
-    yield pytest.param(
-        """
-dbName: src:005:test:rawdb:state
-""",
-        Path("build/raw/raw.yaml"),
-        id="Single Raw Database",
-    )
-
-    yield pytest.param(
-        """
-dbName: src:005:test:rawdb:state
-tableName: myTable
-""",
-        Path("build/raw/raw.yaml"),
-        id="Single Raw Database with table",
-    )
-
-    yield pytest.param(
-        """
-- dbName: src:005:test:rawdb:state
-  tableName: myTable
-- dbName: src:002:weather:rawdb:state
-  tableName: myOtherTable
-""",
-        Path("build/raw/raw.yaml"),
-        id="Multiple Raw Databases with table",
-    )
-
-
-class TestCheckYamlSemantics:
-    @pytest.mark.parametrize("raw_yaml, source_path", list(valid_yaml_semantics_test_cases()))
-    def test_valid_yaml(self, raw_yaml: str, source_path: Path, dummy_environment: Environment) -> None:
-        cmd = BuildCommand(print_warning=False)
-        # Only used in error messages
-        destination = Path("build/raw/raw.yaml")
-        yaml_warnings, *_ = cmd.validate(raw_yaml, source_path, destination)
-        assert not yaml_warnings
-
-    def test_build_valid_read_int_version(self) -> None:
-        cmd = BuildCommand(print_warning=False)
-        destination = Path("build/transformation/transformations.Transformation.yaml")
-        source_path = Path("my_module/transformations/transformations.Transformation.yaml")
-        raw_yaml = """destination:
-  dataModel:
-    destinationType: CogniteFile
-    externalId: MyModel
-    space: my_space
-    version: 1_0_0
-  instanceSpace: my_space
-  type: instances
-externalId: some_external_id
-    """
-        _, identifier_pairs = cmd.validate(raw_yaml, source_path, destination)
-        assert len(identifier_pairs) == 1
-        required_data_model = next(
-            (required_id for loader, required_id in cmd._state.dependencies_by_required if loader is DataModelLoader),
-            None,
-        )
-        assert required_data_model is not None
-        assert required_data_model == DataModelId("my_space", "MyModel", "1_0_0")
